bigint

+++ schema.sql
@@ -44,7 +44,7 @@ CREATE TABLE message (
 );  
 
 CREATE TABLE Files(
-	message_id int not null,
+	message_id bigint not null,
 	file_name Varchar(100) not null,
     file_type varchar(15) not null,
     data blob not null,
